fixed tiny bugs
