+++ src/login.py
@@ -172,17 +172,6 @@ def download():
     ''' Write to disk '''
     pass
 
-def tryload():
-    f = open('tables.json', 'r')
-    data = json.loads(f.read())
-    f.close()
-    glob_dic.tips.set_value('project',data.get('project'))
-    glob_dic.tips.set_value('type',data.get('type'))
-    glob_dic.tips.set_value('issuetype',data.get('issuetype'))
-    glob_dic.tips.set_value('sprint',data.get('sprint'))
-    glob_dic.tips.set_value('status',data.get('status'))
-    # print(data)
-
 # login(['admin','admin'])
 
 # # getStatus()
