+++ src/utils.py
@@ -13,6 +13,8 @@ import logging
 
 import inspect
 
+import json
+
 
 def read_cookie():
     if glob_dic.get_value('cookie') == '' or glob_dic.get_value('cookie') is None:
@@ -126,3 +128,14 @@ def prepare(s):
         'search':{'Accept': 'application/json','cookie':read_cookie()}
     }
     return (address_book.get(s), headers_book.get(s))
+
+def tryload():
+    f = open('tables.json', 'r')
+    data = json.loads(f.read())
+    f.close()
+    glob_dic.tips.set_value('project',data.get('project'))
+    glob_dic.tips.set_value('type',data.get('type'))
+    glob_dic.tips.set_value('issuetype',data.get('issuetype'))
+    glob_dic.tips.set_value('sprint',data.get('sprint'))
+    glob_dic.tips.set_value('status',data.get('status'))
+    # print(data)
